Get torrent downloading working between 2 peers
This is the final step on getting the torrent-stream and peer-wire-swarm modules working with our platform.

What is left to be done is implement a special server component so that peers dont need to maintain extensive memory structures for files they have already downloaded but can still act as seed.